add: corrections in file structure, making of components for the landing page, changed from css to sass
