+++ src/components/animation_button/index.jsx
@@ -1,0 +1,6 @@
+const AnimationButton = () => {
+  return (
+    <button className="registrarse_hero">Regístrarse</button>
+  )
+}
+export default AnimationButton
