+++ src/components/animation_button/styles.scss
@@ -1,0 +1,31 @@
+@import url('https://fonts.googleapis.com/css2?family=Karla:wght@500&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Karla:wght@500;700&family=Montserrat:wght@400;500&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@700&display=swap');
+
+
+.registrarse_hero {
+    background-color: #E91E63;
+    color: white;
+    width: 151px;
+    height: 41px;
+    font-family: 'Karla', sans-serif;
+    font-size: 16px;
+    font-weight: 300;
+    border-radius: 6px;
+    margin: 10px;
+    border: hsl(234,26%,23%) solid 0px;
+    margin-top: 60px;
+    letter-spacing: 0px;
+    transition: all 0.5s;
+}
+
+.registrarse_hero:hover {
+    background-color: white;
+    font-weight: 900;
+    letter-spacing: 2px;
+    border: 3px solid #E91E63;
+    color: #E91E63;
+    border: solid 2px #E91E63;
+    width: 155px;
+    height: 43px;
+}
