+++ src/components/navbar/styles.scss
@@ -1,0 +1,38 @@
+@use '../../styles/app.scss';
+
+
+.navbar.open {
+    height: 50px; /* Adjust the height as needed */
+  }
+
+.navbar {
+    background-color: hsl(234,26%,23%);  
+}
+
+.navbar_icon {
+    font-size: 32px;
+    color:white;
+    cursor: pointer;
+}
+
+.me-auto {
+    font-family: 'Karla', sans-serif;
+}
+
+.navbar-nav {
+    text-align: center;
+    font-size: 18px; 
+}
+
+@media only screen and (min-width: 744px) {
+.desktop_menu {
+    display: flex;
+    align-items:center;
+}
+.navbar-nav {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+}
+
+}
