+++ src/components/pink_button/index.jsx
@@ -1,0 +1,6 @@
+const Pink_button = () => {
+  return (
+    <button className="iniciar_session">Iniciar Sessión</button>
+  )
+}
+export default Pink_button
