+++ src/components/pink_button/styles.css
@@ -1,0 +1,11 @@
+.iniciar_session {
+  background-color: #E91E63;
+  width: 151px;
+  height: 41px;
+  font-family: "Karla", sans-serif;
+  font-size: 16px;
+  border: none;
+  color: white;
+  border-radius: 6px;
+  margin: 10px;
+}/*# sourceMappingURL=styles.css.map */
