+++ src/components/pink_button/styles.scss
@@ -1,0 +1,11 @@
+.iniciar_session {
+    background-color: #E91E63;
+    width: 151px;
+    height: 41px;
+    font-family: 'Karla', sans-serif;
+    font-size: 16px;
+    border: none;
+    color: white;
+    border-radius: 6px;
+    margin: 10px;
+}
